+++ compile-shaders.sh
@@ -112,6 +112,7 @@ compile pick                         pick.comp             vulkan1.0
 compile conv                         conv.comp             vulkan1.0
 compile rule                         rule.comp             vulkan1.0
 compile rule_held                    rule_held.comp        vulkan1.1
+compile rule_held_single             rule_held.comp        vulkan1.1 SINGLE
 
 # shellcheck disable=SC2086
 (cd "$ROOT/tests" && "$TESTS" shader $PAIRS)